requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
matplotlib>=3.7.0
SQLAlchemy>=2.0.0
GeoAlchemy2>=0.14.0

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-telegram-bot>=20.7
langchain>=0.2.14,<0.3.0
langchain-community>=0.2.12,<0.3.0
langchain-core>=0.2.35,<0.3.0
langchain-openai>=0.1.22,<0.2.0
openai>=1.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

qdrant-client==1.15.1
sentence-transformers>=2.3.0

redis>=5.0.1
aioredis>=2.0.1

PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3

arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.25
openinference-instrumentation-openai>=0.1.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

